feat(aliketeller): extract logic from server and put on context and client
